Add file param to stdout, richoutput, rm stderr
This patch adds a file parameter to the StdoutLog, defaulting to
sys.stdout, and removes the StderrLog in favour of a StdoutLog with
file=sys.stderr. The expectation is the StderrLog is never used and that
it can be removed without warning.

The unit tests are changed because the capture function relied on
contextlib.redirect_stdout and contextlib.redirect_stderr, which no
longer has effect if the logger has a handle to the file object. The new
strategy is to create a StdoutLog with a StringIO argument and read out
the value afterward. This has the important advantage of of leaving
sys.stdout intact so that error messages can be read if a test fails.